requirements.txt
python-fasthtml
ultralytics
pillow
python-multipart